2018-11-21-1055CST Version, Updated 2018-11-21-1310CST
2018-11-21-1055CST VERSION, Updated 2018-11-21-1310CST
Serious revamp on some scripts with efforts to reduce impact from Gaia clish calls.  Updated templates.  Session_Cleanup now based on API script template v00.32.00 and fully capable of CLI configuration for all mgmt_cli calls.  Healthcheck v 6.0 script.

+++ scripts/updatescripts.sh
@@ -5,11 +5,11 @@
 # (C) 2016-2018 Eric James Beasley, @mybasementcloud, https://github.com/mybasementcloud/bash_4_Check_Point_scripts
 #
 ScriptTemplateLevel=005
-ScriptVersion=01.00.00
-ScriptDate=2018-10-25
+ScriptVersion=02.00.00
+ScriptDate=2018-11-21
 #
 
-export BASHScriptVersion=v01x00x00
+export BASHScriptVersion=v02x00x00
 export BASHScriptTemplateLevel=$ScriptTemplateLevel
 export BASHScriptName="updatescripts"
 export BASHScriptShortName="updatescripts"
@@ -24,8 +24,9 @@ export DATEDTG=`date +%Y-%m-%d-%H%M%Z`
 export DATEDTGS=`date +%Y-%m-%d-%H%M%S%Z`
 export DATEYMD=`date +%Y-%m-%d`
 
-WAITTIME=60
+WAITTIME=20
 
+targettftpserver=10.69.248.60
 targetfolder=/var/log/__customer/upgrade_export
 
 # setup initial log file for output logging
@@ -61,9 +62,10 @@ echo
 read -t $WAITTIME -n 1 -p "Any key to continue.  Automatic continue after $WAITTIME seconds : " anykey
 echo
 
-echo "Fetch scripts.tgz from tftp repository..." | tee -a -i $logfilepath
+echo "Fetch latest scripts.tgz from tftp repository on $targettftpserver..." | tee -a -i $logfilepath
 echo | tee -a -i $logfilepath
-tftp -v -m binary 10.69.248.60 -c get /__gaia/scripts.tgz | tee -a -i $logfilepath
+tftp -v -m binary $targettftpserver -c get /__gaia/scripts.tgz | tee -a -i $logfilepath
+tftp -v -m binary $targettftpserver -c get /__gaia/updatescripts.sh | tee -a -i $logfilepath
 echo | tee -a -i $logfilepath
 
 echo "Check that we got it." | tee -a -i $logfilepath
@@ -77,18 +79,57 @@ else
     # we have the scripts.tgz file and can work with it
     # first remove the existing script links
     
-    echo "Remove existing script links..." | tee -a -i $logfilepath
-    echo | tee -a -i $logfilepath
-    export removefile=`ls ./scripts/remove_script*`
-    . $removefile | tee -a -i $logfilepath
-    echo | tee -a -i $logfilepath
+    if [ -r $targetfolder/scripts ]; then    
+        # scripts folder exists, so we might have scripts
+
+        export removefile=`ls $targetfolder/scripts/remove_script*`
+        if [ -r $removefile ]; then    
+            echo "Remove existing script links... $removefile" | tee -a -i $logfilepath
+            echo | tee -a -i $logfilepath
+            . $removefile | tee -a -i $logfilepath
+            echo | tee -a -i $logfilepath
+        fi
+        
+        # next remove existing scripts folder
+        echo "Remove scripts folder... $targetfolder/scripts" | tee -a -i $logfilepath
+        echo | tee -a -i $logfilepath
+        
+        rm -f -r -d $targetfolder/scripts | tee -a -i $logfilepath
+
+        # next [re-]create scripts folder
+        echo "Create/Check scripts folder... $targetfolder/scripts" | tee -a -i $logfilepath
+        echo | tee -a -i $logfilepath
+        
+        if [ -r $targetfolder/scripts ]; then
+            mkdir $targetfolder/scripts | tee -a -i $logfilepath
+            chmod 775 $targetfolder/scripts | tee -a -i $logfilepath
+        else
+            chmod 775 $targetfolder/scripts | tee -a -i $logfilepath
+        fi
     
-    # next remove existing scripts folder
-    echo "Remove scripts folder..." | tee -a -i $logfilepath
+        export oldversionfile=`ls $targetfolder/scripts.*.version`
+        if [ -r $oldversionfile ]; then    
+            echo "Remove old version file from main target folder... $oldversionfile" | tee -a -i $logfilepath
+            echo | tee -a -i $logfilepath
+            rm -f $oldversionfile | tee -a -i $logfilepath
+            echo | tee -a -i $logfilepath
+        fi
+    else
+        # scripts folder doesn't exists, so we might have scripts
+
+        # next [re-]create scripts folder
+        echo "Create/Check scripts folder..." | tee -a -i $logfilepath
+        echo | tee -a -i $logfilepath
+        
+        if [ -r $targetfolder/scripts ]; then
+            mkdir $targetfolder/scripts | tee -a -i $logfilepath
+            chmod 775 $targetfolder/scripts | tee -a -i $logfilepath
+        else
+            chmod 775 $targetfolder/scripts | tee -a -i $logfilepath
+        fi
+    fi
     echo | tee -a -i $logfilepath
-    
-    rm -f -r -d $targetfolder/scripts | tee -a -i $logfilepath
-    
+
     # now unzip existing scripts folder
     echo "Extract scripts.tgz file..." | tee -a -i $logfilepath
     echo | tee -a -i $logfilepath
@@ -100,13 +141,20 @@ else
     ls -alhR ./scripts | tee -a -i $logfilepath
     echo | tee -a -i $logfilepath
     
-    echo "Generate script links..." | tee -a -i $logfilepath
-    echo | tee -a -i $logfilepath
-
     export generatefile=`ls ./scripts/generate_script*`
+    echo "Generate script links... $generatefile" | tee -a -i $logfilepath
+    echo | tee -a -i $logfilepath
     . $generatefile | tee -a -i $logfilepath
     echo | tee -a -i $logfilepath
     
+    export versionfile=`ls $targetfolder/scripts/scripts.*.version`
+    if [ -r $versionfile ]; then    
+        echo "Copy version file to main target folder... $versionfile" | tee -a -i $logfilepath
+        echo | tee -a -i $logfilepath
+        cp $versionfile . | tee -a -i $logfilepath
+        echo | tee -a -i $logfilepath
+    fi
+
     echo 'Done!' | tee -a -i $logfilepath
     echo | tee -a -i $logfilepath
 fi
